Moving Entity.__init__ to static.py to avoid overwrite by fetch_schema
Has implications on the placement of methods and classes in osw.utils.strings, as this module imports osw.model.static.Basemodel - we need to avoid circular imports
