+++ scripts/migration/file_page_migration.py
@@ -12,8 +12,8 @@ import osw.data.import_utility as iu
 import osw.wiki_tools as wt
 from osw.core import OSW
 from osw.model.entity import Label, WikiFile
+from osw.utils.regex import match_first_regex_pattern, test_regex_pattern
 from osw.utils.regex_pattern import REGEX_PATTERN_LIB, REGEX_PATTERN_LIST
-from osw.utils.strings import match_first_regex_pattern, test_regex_pattern
 from osw.utils.util import parallelize
 from osw.wtsite import WtPage, WtSite
