+++ src/osw/controller/page_package.py
@@ -14,7 +14,7 @@ from osw.auth import CredentialManager
 from osw.model import page_package as package
 from osw.model.page_package import NAMESPACE_CONST_TO_NAMESPACE_MAPPING
 from osw.model.static import OswBaseModel
-from osw.utils.strings import RegExPatternExtended
+from osw.utils.regex import RegExPatternExtended
 from osw.wtsite import WtSite
 
 # Definition of constants
